pyproject.toml
[project]
name = "atmacup18"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "albumentations>=1.4.21",
    "catboost>=1.2.7",
    "jupyterlab>=4.3.0",
    "lightgbm>=4.5.0",
    "matplotlib>=3.9.2",
    "numpy>=1.26.4",
    "opencv-python-headless>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=5.24.1",
    "polars>=1.13.1",
    "pyarrow>=18.0.0",
    "pytorch-lightning>=2.4.0",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "timm>=1.0.11",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.0",
    "transformers>=4.46.2",
    "wandb>=0.18.7",
    "xgboost>=2.1.2",
]

[dependency-groups]
dev = [
    "isort>=5.13.2",
    "pydantic>=2.9.2",
    "ruff>=0.7.3",
]
